backend/requirements.txt
fastapi
uvicorn
rembg
python-multipart
Pillow
numpy
